# docs/swagger.yaml
basePath: /api
definitions:
  models.AdminUser:
    properties:
      created_at:
        type: string
      created_by:
        type: integer
      creator:
        allOf:
        - $ref: '#/definitions/models.AdminUser'
        description: Relations
      email:
        type: string
      id:
        type: integer
      last_login_at:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      role:
        enum:
        - super_admin
        - admin
        - operator
        - moderator
        type: string
      status:
        enum:
        - active
        - inactive
        type: string
      updated_at:
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - name
    - username
    type: object
  models.Category:
    properties:
      created_at:
        type: string
      description:
        type: string
      display_order:
        type: integer
      icon_url:
        type: string
      id:
        type: integer
      name:
        maxLength: 100
        minLength: 2
        type: string
      products:
        description: Relations
        items:
          $ref: '#/definitions/models.Product'
        type: array
      slug:
        type: string
      status:
        enum:
        - active
        - inactive
        type: string
      updated_at:
        type: string
    required:
    - name
    type: object
  models.FormField:
    properties:
      field:
        type: string
      label:
        type: string
      options:
        items:
          type: string
        type: array
      placeholder:
        type: string
      required:
        type: boolean
      type:
        enum:
        - text
        - select
        - number
        type: string
    required:
    - field
    - label
    - type
    type: object
  models.GameAccount:
    properties:
      created_at:
        type: string
      game_id:
        maxLength: 100
        type: string
      game_name:
        maxLength: 100
        type: string
      id:
        type: integer
      is_primary:
        type: boolean
      nickname:
        maxLength: 100
        type: string
      server:
        maxLength: 50
        type: string
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relations
      user_id:
        type: integer
      zone_id:
        maxLength: 50
        type: string
    required:
    - game_id
    - game_name
    - user_id
    type: object
  models.GameAccountData:
    properties:
      game_account:
        type: string
      game_server:
        type: string
      game_zone:
        type: string
      nickname:
        type: string
    required:
    - game_account
    type: object
  models.Product:
    properties:
      category:
        allOf:
        - $ref: '#/definitions/models.Category'
        description: Relations
      category_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      discount_percentage:
        maximum: 100
        minimum: 0
        type: integer
      display_order:
        type: integer
      form_config:
        items:
          $ref: '#/definitions/models.FormField'
        type: array
      id:
        type: integer
      image_url:
        type: string
      name:
        maxLength: 200
        minLength: 2
        type: string
      original_price:
        minimum: 0
        type: number
      price:
        minimum: 0
        type: number
      slug:
        type: string
      status:
        enum:
        - active
        - inactive
        - out_of_stock
        type: string
      transactions:
        items:
          $ref: '#/definitions/models.Transaction'
        type: array
      updated_at:
        type: string
    required:
    - category_id
    - form_config
    - name
    - price
    type: object
  models.Transaction:
    properties:
      completed_at:
        type: string
      created_at:
        type: string
      expired_at:
        type: string
      game_account_data:
        allOf:
        - $ref: '#/definitions/models.GameAccountData'
        description: Game Account Info (denormalized for transaction history)
      id:
        type: integer
      ip_address:
        type: string
      logs:
        items:
          $ref: '#/definitions/models.TransactionLog'
        type: array
      payment_fee:
        minimum: 0
        type: number
      payment_method:
        description: Payment
        enum:
        - gopay
        - ovo
        - dana
        - bca
        - mandiri
        - bni
        type: string
      payment_reference:
        type: string
      payment_status:
        enum:
        - pending
        - paid
        - failed
        - expired
        - refunded
        type: string
      payment_url:
        type: string
      processed_at:
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: integer
      product_price:
        description: Pricing
        minimum: 0
        type: number
      status:
        description: Status & Processing
        enum:
        - pending
        - processing
        - completed
        - failed
        - cancelled
        type: string
      total_amount:
        minimum: 0
        type: number
      transaction_code:
        type: string
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relations
      user_agent:
        description: Metadata
        type: string
      user_id:
        type: integer
      voucher_discount:
        minimum: 0
        type: number
      voucher_usages:
        items:
          $ref: '#/definitions/models.VoucherUsage'
        type: array
      whatsapp:
        description: Contact
        type: string
    required:
    - game_account_data
    - payment_method
    - product_id
    - product_price
    - total_amount
    - user_id
    - whatsapp
    type: object
  models.TransactionLog:
    properties:
      admin:
        $ref: '#/definitions/models.AdminUser'
      created_at:
        type: string
      created_by_admin:
        type: integer
      id:
        type: integer
      message:
        type: string
      metadata:
        additionalProperties: true
        type: object
      status_from:
        type: string
      status_to:
        type: string
      transaction:
        allOf:
        - $ref: '#/definitions/models.Transaction'
        description: Relations
      transaction_id:
        type: integer
    required:
    - status_to
    - transaction_id
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      email_verified_at:
        type: string
      game_accounts:
        description: Relations
        items:
          $ref: '#/definitions/models.GameAccount'
        type: array
      id:
        type: integer
      name:
        maxLength: 100
        minLength: 2
        type: string
      phone:
        type: string
      phone_verified_at:
        type: string
      status:
        enum:
        - active
        - inactive
        - suspended
        type: string
      transactions:
        items:
          $ref: '#/definitions/models.Transaction'
        type: array
      updated_at:
        type: string
    required:
    - name
    - phone
    type: object
  models.Voucher:
    properties:
      application_type:
        enum:
        - all
        - category
        - product
        type: string
      applications:
        description: Relations
        items:
          $ref: '#/definitions/models.VoucherApplication'
        type: array
      code:
        maxLength: 50
        minLength: 3
        type: string
      created_at:
        type: string
      description:
        type: string
      end_date:
        type: string
      id:
        type: integer
      max_discount_amount:
        minimum: 0
        type: number
      max_uses_per_user:
        minimum: 1
        type: integer
      min_transaction_amount:
        minimum: 0
        type: number
      quota:
        minimum: 1
        type: integer
      start_date:
        type: string
      status:
        enum:
        - active
        - inactive
        - expired
        type: string
      type:
        enum:
        - percentage
        - fixed
        type: string
      updated_at:
        type: string
      usages:
        items:
          $ref: '#/definitions/models.VoucherUsage'
        type: array
      used_count:
        type: integer
      value:
        minimum: 0
        type: number
    required:
    - code
    - end_date
    - quota
    - start_date
    - type
    - value
    type: object
  models.VoucherApplication:
    properties:
      applicable_id:
        type: integer
      applicable_type:
        enum:
        - category
        - product
        type: string
      created_at:
        type: string
      id:
        type: integer
      voucher:
        allOf:
        - $ref: '#/definitions/models.Voucher'
        description: Relations
      voucher_id:
        type: integer
    required:
    - applicable_id
    - applicable_type
    - voucher_id
    type: object
  models.VoucherUsage:
    properties:
      discount_amount:
        minimum: 0
        type: number
      id:
        type: integer
      transaction:
        $ref: '#/definitions/models.Transaction'
      transaction_id:
        type: integer
      used_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
      voucher:
        allOf:
        - $ref: '#/definitions/models.Voucher'
        description: Relations
      voucher_id:
        type: integer
    required:
    - discount_amount
    - transaction_id
    - user_id
    - voucher_id
    type: object
...
  services.CreateCategoryRequest:
    properties:
      description:
        type: string
      display_order:
        type: integer
      name:
        maxLength: 100
        minLength: 2
        type: string
    required:
    - name
    type: object
...
  services.UpdateCategoryRequest:
    properties:
      description:
        type: string
      display_order:
        type: integer
      name:
        maxLength: 100
        minLength: 2
        type: string
      status:
        enum:
        - active
        - inactive
        type: string
    required:
    - name
    - status
    type: object
...
  utils.APIResponse:
    properties:
      data: {}
      errors:
        items:
          type: string
        type: array
      message:
        type: string
      pagination:
        $ref: '#/definitions/utils.Pagination'
      success:
        type: boolean
    type: object
  utils.Pagination:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@wawstore.com
    name: API Support
    url: https://wawstore.com/support
  description: Backend API for Waw Store - Game Topup Platform
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://wawstore.com/terms
  title: Waw Store Topup Game Online API
  version: "1.0"
paths:
...
  /admin/categories:
    get:
      description: Get paginated categories for admin management
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Category'
                  type: array
                pagination:
                  $ref: '#/definitions/utils.Pagination'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get categories for admin
      tags:
      - admin
    post:
      consumes:
      - application/json
      description: Create a new category (Admin only)
      parameters:
      - description: Category creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Category'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Create a new category
      tags:
      - admin
  /admin/categories/{id}:
    delete:
      description: Delete a category by ID (Admin only)
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete a category
      tags:
      - admin
    put:
      consumes:
      - application/json
      description: Update a category by ID (Admin only)
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Category update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Category'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Update a category
      tags:
      - admin
...
  /categories:
    get:
      description: Get all active categories for public view
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Category'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Get all categories
      tags:
      - categories
  /categories/{id}:
    get:
      description: Get a specific category by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Category'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Get category by ID
      tags:
      - categories
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"